+++ src/lib/career/index.ts
@@ -5,4 +5,3 @@
 export * from './branchTree';
 export * from './validation';
 export * from './colorAssignment';
-export * from './layoutCalculation';
